backend/requirements.txt
django==2.2.16
django-filter==22.1
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.4.0
psycopg2-binary==2.8.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
Pillow==9.1.1
requests==2.26.0
reportlab==3.6.9